- allow upgrading on missing crontabs
Changed files:
    eventum.spec -> 1.316

+++ eventum.spec
@@ -10,8 +10,8 @@
 %bcond_without	order	# with experimental order patch
 
 #define	snap	20060921
-%define	rev		r4075
-%define	rel		2.62
+%define	rev		r4089
+%define	rel		2.64
 #define	_rc		RC3
 
 %define		php_min_version 5.1.2
@@ -28,7 +28,7 @@ Group:		Applications/WWW
 #Source0:	http://mysql.easynet.be/Downloads/eventum/%{name}-%{version}.tar.gz
 # bzr branch lp:eventum eventum && cd eventum && make dist
 Source0:	%{name}-%{version}-dev-%{rev}.tar.gz
-# Source0-md5:	6fe28279980f06136ae38e94d907549d
+# Source0-md5:	8dbcfe39d29efa4cb8b2dd8c230fe979
 Source1:	%{name}-apache.conf
 Source2:	%{name}-mail-queue.cron
 Source3:	%{name}-mail-download.cron
@@ -663,6 +663,7 @@ chgrp http %{_webappdir}/setup.php
 chgrp http /var/log/%{name}/*
 # update crontab user
 for a in /etc/cron.d/eventum-*; do
+	[ -f "$a" ] || continue
 	awk '!/#/ && NR > 6 && $6 =="eventum" {sub("eventum", "http", $6)}{print}'  $a > $a.rpmtmp && cat $a.rpmtmp > $a
 	rm -f $a.rpmtmp
 done
